Rename attribute_registry to input_registry
