+++ lib/inspec.rb
@@ -16,7 +16,7 @@ require 'inspec/runner'
 require 'inspec/shell'
 require 'inspec/formatters'
 require 'inspec/reporters'
-require 'inspec/attribute_registry'
+require 'inspec/input_registry'
 require 'inspec/rspec_extensions'
 require 'inspec/globals'
 require 'inspec/impact'
