+++ lib/inspec/input_registry.rb
